estoque_pdv.quantidade passa a NUMERIC: saldo fracionario nao trava mais a assinatura do inventario
- Migracao tools/migrar-estoque-pdv-numeric.mjs (aplicada; totais conferidos)
- Remove arredondamento do saldo em SALDOS (produto em KG/ML nao pode ser arredondado)
- Leituras de saldo do PDV sem asInt/::int (index.js e avarias)
- Dump de estrutura regenerado e validado (10/10)

+++ db/estrutura.sql
@@ -563,7 +563,7 @@ CREATE TABLE public.estoque_pdv (
     id integer NOT NULL,
     pdv_id integer,
     sku_produto text,
-    quantidade integer DEFAULT 0,
+    quantidade numeric DEFAULT 0,
     estoque_minimo integer DEFAULT 0,
     estoque_maximo integer DEFAULT 0,
     permitido boolean DEFAULT false,
